+++ src/inspect_ai/_eval/eval.py
@@ -23,7 +23,7 @@ from inspect_ai.model import (
     GenerateConfigArgs,
     Model,
 )
-from inspect_ai.model._model import resolve_models
+from inspect_ai.model._model import init_active_model, resolve_models
 from inspect_ai.solver import Plan, Solver
 from inspect_ai.tool import ToolEnvironmentSpec
 from inspect_ai.tool._environment.context import startup_tool_environments
@@ -223,9 +223,11 @@ async def eval_async(
             model, model_base_url, model_args, GenerateConfig(**kwargs)
         )
 
-        # resolve tasks
+        # resolve tasks (set active model to resolve uses of the
+        # 'default' model in tools, solvers, and scorers)
         resolved_tasks: list[ResolvedTask] = []
         for m in models:
+            init_active_model(m)
             resolved_tasks.extend(resolve_tasks(tasks, task_args, m, toolenv))
 
         # warn and return empty string if we resolved no tasks